Move aws proxy from xray/internal to internal/aws.
