+++ internal/aws/proxy/Makefile
@@ -1,0 +1,1 @@
+include ../../../Makefile.Common
