+++ internal/aws/proxy/go.mod
@@ -1,0 +1,10 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/internal/aws/proxy
+
+go 1.16
+
+require (
+	github.com/aws/aws-sdk-go v1.40.19
+	github.com/stretchr/testify v1.7.0
+	go.opentelemetry.io/collector v0.31.1-0.20210810171211-8038673eba9e
+	go.uber.org/zap v1.19.0
+)
