+++ internal/aws/proxy/testdata/ecsmetadatafile.txt
