+++ internal/aws/proxy/testdata/ecsmetadatafileInvalidArn.txt
