+++ internal/aws/proxy/testdata/ecsmetadatafileinvalid.txt
